+++ CHANGES.md
@@ -22,6 +22,7 @@
 -   Made CSW connector gracefully handle datasets without ids instead of crashing
 -   Fixed Gitlab UI only preview failed to download main CSS file
 -   Removed CORS handling from Scala APIs, should be entirely handled by the gateway.
+-   Allows gateway to redirect trailing slash for APIDocs module
 
 ## 0.0.48